[Core] Extract field value helper declarations
* Moved generated field helper overloads into field_values.hpp

* Simplified numeric field helper templates with shared concepts
